payment-service/requirements.txt
fastapi
uvicorn[standard]
mongoengine
python-jose[cryptography]
requests
python-dotenv
pydantic
resend
mangum